Split behavior method restoration tests
